+++ packages/gamut/src/Form/FormError.tsx
@@ -3,18 +3,19 @@ import { StyleProps } from '@codecademy/variance';
 import styled from '@emotion/styled';
 import React, { HTMLAttributes } from 'react';
 
-import { colorStates } from './styles/shared';
-
 const errorSpanVariants = variant({
   base: {
     left: 0,
     top: `calc(100% - ${theme.spacing[8]})`,
-    color: colorStates.error.color,
+    color: 'feedback-error',
     display: 'inline-block',
     fontSize: 'small',
   },
-  defaultVariant: 'absolute',
+  defaultVariant: 'initial',
   variants: {
+    initial: {
+      position: 'initial',
+    },
     absolute: {
       position: 'absolute',
     },