+++ packages/gamut/src/GridForm/GridFormInputGroup/GridFormCheckboxInput/index.tsx
@@ -6,7 +6,6 @@ import { GridFormCheckboxField } from '../../types';
 
 export type GridFormCheckboxInputProps = {
   className?: string;
-  clearErrors: UseFormMethods['clearErrors'];
   field: GridFormCheckboxField;
   register: UseFormMethods['register'];
   showRequired?: boolean;
@@ -14,7 +13,6 @@ export type GridFormCheckboxInputProps = {
 
 export const GridFormCheckboxInput: React.FC<GridFormCheckboxInputProps> = ({
   className,
-  clearErrors,
   field,
   register,
   showRequired,
@@ -26,10 +24,7 @@ export const GridFormCheckboxInput: React.FC<GridFormCheckboxInputProps> = ({
       disabled={field.disabled}
       htmlFor={field.name}
       name={field.name}
-      onChange={(event) => {
-        clearErrors();
-        field.onUpdate?.(event.target.checked);
-      }}
+      onChange={(event) => field.onUpdate?.(event.target.checked)}
       label={field.description}
       multiline={field.multiline}
       ref={register(field.validation)}